Extract logging logic to separate file
